Move cgroup and oom score setting to cmd.
