+++ cmd/cri-containerd/cri_containerd.go
@@ -27,6 +27,9 @@ import (
 	"runtime"
 	"syscall"
 
+	"github.com/containerd/cgroups"
+	"github.com/containerd/containerd/sys"
+	runtimespec "github.com/opencontainers/runtime-spec/specs-go"
 	"github.com/opencontainers/selinux/go-selinux"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
@@ -94,6 +97,19 @@ func main() {
 		}
 
 		logrus.Infof("Run cri-containerd %+v", o)
+
+		if o.CgroupPath != "" {
+			_, err := loadCgroup(o.CgroupPath)
+			if err != nil {
+				return fmt.Errorf("failed to load cgroup for cgroup path %v: %v", o.CgroupPath, err)
+			}
+		}
+
+		if o.OOMScore != 0 {
+			if err := sys.SetOOMScore(os.Getpid(), o.OOMScore); err != nil {
+				return fmt.Errorf("failed to set OOMScore to %v: %v", o.OOMScore, err)
+			}
+		}
 
 		// Start profiling server if enable.
 		if o.EnableProfiling {
@@ -199,4 +215,24 @@ func setGLogLevel(l logrus.Level) error {
 	case logrus.PanicLevel:
 	}
 	return nil
+}
+
+// loadCgroup loads the cgroup associated with path if it exists and moves the current process into the cgroup. If the cgroup
+// is not created it is created and returned.
+func loadCgroup(cgroupPath string) (cgroups.Cgroup, error) {
+	cg, err := cgroups.Load(cgroups.V1, cgroups.StaticPath(cgroupPath))
+	if err != nil {
+		if err != cgroups.ErrCgroupDeleted {
+			return nil, err
+		}
+		if cg, err = cgroups.New(cgroups.V1, cgroups.StaticPath(cgroupPath), &runtimespec.LinuxResources{}); err != nil {
+			return nil, err
+		}
+	}
+	if err := cg.Add(cgroups.Process{
+		Pid: os.Getpid(),
+	}); err != nil {
+		return nil, err
+	}
+	return cg, nil
 }
